1.3
Modificados los schemas, el archivo auth.js, el hepers para añadirle el path, cambiada la base de datos para poner añadir una imagen a la tabla de comentarios. Añadidas funciones para los comentarios, y ya solo puede borrar los comentarios el autor del comentario

+++ controllers/comments.js
@@ -3,7 +3,14 @@ const {
   newCommentSchema,
   idCommentsSchema,
 } = require('../schemas/commentsSchemas');
-const { createComments } = require('../db/comments.js');
+const {
+  createComments,
+  getCommentsByRecommendations,
+  getCommentById,
+} = require('../db/comments.js');
+const path = require('path');
+const {createPathIfNotExists} = require ('../helpers')
+const sharp = require('sharp')
 
 const newCommentController = async (req, res, next) => {
   try {
@@ -12,8 +19,24 @@ const newCommentController = async (req, res, next) => {
       res.status(400).json({ error: error.details[0].message });
       return;
     }
+    let imageFileName;
+    if (req.files && req.files.image){
+  const uploadsDir = path.join(__dirname, '../uploads');
+  await createPathIfNotExists(uploadsDir)
+  const { default: nanoid } = await import('nanoid');
+  const image = sharp(req.files.image.data)
+  imageFileName = `${nanoid(20).jpg}`
+  await image.toFile(path.join(uploadsDir, imageFileName))
+
+    }
+    
     const { user_id, recommendation_id, comment } = req.body;
-    const commentId = await createComments(user_id, recommendation_id, comment);
+    const commentId = await createComments(
+      user_id,
+      recommendation_id,
+      comment,
+      imageFileName
+    );
     return res
       .status(201)
       .json({ message: 'Comment posted successfully', commentId });
@@ -28,13 +51,8 @@ const getCommentsByRecommendationsController = async (req, res, next) => {
     if (error) {
       return res.status(400).json({ error: 'Invalid id' });
     }
-    const recommendationId = req.params.id;
-    const connection = await getConnection();
-    const [rows] = await connection.query(
-      'SELECT comments.*, users.name as username FROM comments INNER JOIN users ON comments.user_id = users.id WHERE recommendation_id = ?',
-      [recommendationId]
-    );
-    connection.release();
+
+    const comments = await getCommentsByRecommendations(req, res);
 
     res.status(200).json({ comments: rows });
   } catch (err) {
@@ -44,21 +62,31 @@ const getCommentsByRecommendationsController = async (req, res, next) => {
 };
 
 const deleteCommentsByUserController = async (req, res, next) => {
+  let connection;
   try {
     const { error } = idCommentsSchema.validate(req.params);
     if (error) {
       return res.status(400).json({ error: 'Invalid id' });
     }
-    const commentId = req.params.id;
-    const connection = await getConnection();
+    const { id } = req.params;
+    const comment = await getCommentById(id);
+    console.log(req.userId);
+    console.log(comment.user_id);
+    if (req.userId !== comment.user_id) {
+      throw new Error(
+        'You are trying to delete a comment that does not belong to you'
+      );
+    }
+    connection = await getConnection();
     const deleteCommentQuery = 'DELETE FROM comments WHERE id = ?';
-    await connection.query(deleteCommentQuery, [commentId]);
-    connection.release();
+    await connection.query(deleteCommentQuery, [id]);
 
     res.status(200).json({ message: 'Comment deleted successfully' });
   } catch (err) {
     console.error(err);
     res.status(500).json({ error: 'Server error' });
+  } finally {
+    if (connection) connection.release();
   }
 };
